refactor: header 폴더 생성 및 각 컴포넌트의 최상위 파일명 변경
- index.js가 각 컴포넌트의 최상위 파일
