+++ src/js/components/mainContent/index.js
